+++ www/src/types/plotly.js-cartesian-dist-min.d.ts
@@ -1,4 +1,4 @@
-declare module "plotly.js-dist-min" {
+declare module "plotly.js-cartesian-dist-min" {
   const Plotly: typeof import("plotly.js");
   export default Plotly;
 }